requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
python-decouple==3.8
sqlparse==0.5.3
coreapi==2.3.3
drf-yasg==1.21.10
django-mail-templated==2.6.5
black==24.8.0
flake8==7.0.0
pytest==8.4.1 
pytest-django==4.11.1